Moved election summary fixtures based on Marks comment and discussion
