+++ backend/src/summary/mod.rs
@@ -224,77 +224,10 @@ impl SumCount {
 }
 
 #[cfg(test)]
-pub(crate) mod tests {
+mod tests {
     use super::*;
-    use crate::{
-        election::{PGNumber, tests::election_fixture},
-        pdf_gen::tests::polling_stations_fixture,
-    };
+    use crate::{election::tests::election_fixture, pdf_gen::tests::polling_stations_fixture};
     use test_log::test;
-
-    /// Create a test election summary with given total votes and political group votes.
-    pub fn election_summary_fixture_given_political_group_votes(
-        total_votes: Count,
-        political_group_votes: Vec<PoliticalGroupVotes>,
-    ) -> ElectionSummary {
-        ElectionSummary {
-            voters_counts: VotersCounts {
-                poll_card_count: total_votes,
-                proxy_certificate_count: 0,
-                voter_card_count: 0,
-                total_admitted_voters_count: total_votes,
-            },
-            votes_counts: VotesCounts {
-                votes_candidates_count: total_votes,
-                blank_votes_count: 0,
-                invalid_votes_count: 0,
-                total_votes_cast_count: total_votes,
-            },
-            differences_counts: SummaryDifferencesCounts::zero(),
-            recounted_polling_stations: vec![],
-            political_group_votes,
-        }
-    }
-
-    /// Create a test election summary with given votes per political group.  
-    /// The number of political groups is the length of the `pg_votes` vector.  
-    /// The number of candidates in each political group is by default 50.
-    pub fn election_summary_fixture_with_default_50_candidates(
-        pg_votes: Vec<Count>,
-    ) -> ElectionSummary {
-        let total_votes = pg_votes.iter().sum();
-        let mut political_group_votes: Vec<PoliticalGroupVotes> = vec![];
-        for (index, votes) in pg_votes.iter().enumerate() {
-            // Create list with 50 candidates with 0 votes
-            let mut candidate_votes: Vec<Count> = vec![0; 50];
-            // Set votes to first candidate
-            candidate_votes[0] = *votes;
-            political_group_votes.push(PoliticalGroupVotes::from_test_data_auto(
-                PGNumber::try_from(index + 1).unwrap(),
-                *votes,
-                &candidate_votes,
-            ))
-        }
-        election_summary_fixture_given_political_group_votes(total_votes, political_group_votes)
-    }
-
-    /// Create a test election summary with given candidate votes per political group.  
-    /// The number of political groups is the length of the `candidate_votes` vector.  
-    /// The number of candidates in each political group is equal to the value in the vector at that index.
-    pub fn election_summary_fixture_with_given_candidate_votes(
-        candidate_votes: Vec<Vec<Count>>,
-    ) -> ElectionSummary {
-        let total_votes = candidate_votes.iter().flatten().sum();
-        let mut political_group_votes: Vec<PoliticalGroupVotes> = vec![];
-        for (pg_index, pg_candidate_votes) in candidate_votes.iter().enumerate() {
-            political_group_votes.push(PoliticalGroupVotes::from_test_data_auto(
-                PGNumber::try_from(pg_index + 1).unwrap(),
-                pg_candidate_votes.iter().sum(),
-                pg_candidate_votes,
-            ))
-        }
-        election_summary_fixture_given_political_group_votes(total_votes, political_group_votes)
-    }
 
     fn polling_station_results_fixture_a() -> PollingStationResults {
         PollingStationResults {
